add, update and register tests

+++ integration_tests/where_01.f90
@@ -1,6 +1,6 @@
 program where_01
 implicit none
-real :: a(10), b(10)
+integer :: a(10), b(10)
 where (a >= 0)
     b = 1
 else where
